services/analyzer/requirements.txt
celery>=5.3
pyyaml>=6.0
requests>=2.32
lancedb>=0.8
pyarrow>=15.0
pdf2image>=1.17.0
python-Levenshtein>=0.25.1
python-dotenv>=1.0.1
